EIO Utilities, Shooter Limit, Tuning Command
Added an ArgumentProvider class, an easy way to wrap a variable for
later use.
Added an Analog Argument Provider, which wraps the ArgumentProvider
class to provide an analog value from the driver station.
Added an AnalogChangeTrigger, a button wrapper that triggers when it
sees a big enough change in the analog value being read from the driver
station.
Added the shooter limit switch implementation.
Changed the setPitchMotorSpeed() function of shooter to take direction
instead of actual speed.