Command waiting and async collector
Provided a setting in the change shooter pitch command to wait for the
collector to be out of the way before running, instead of just not
running at all.
Changed the last steps of the CollectorShooterLoad group to run
asynchronous to each other; the shooter and the collector will both run
in parallel.